Generate documentation with Redoc CLI

+++ src/Console/Generate.php
@@ -7,19 +7,47 @@ use mikehaertl\shellcommand\Command;
 
 class Generate extends \Illuminate\Console\Command
 {
+    const GENERATOR_REDOC = 'redoc';
+
     protected $signature = 'openapi:generate {--definition=main} {generator} {output}';
+
+    private function usingOpenApiGenerator(): bool
+    {
+        return !$this->usingRedoc();
+    }
+
+    private function usingRedoc(): bool
+    {
+        return ($this->argument('generator') === self::GENERATOR_REDOC);
+    }
 
     public function handle()
     {
-        if (!$this->openApiGeneratorCommandExists()) {
+        if ($this->usingOpenApiGenerator() && !$this->openApiGeneratorCommandExists()) {
             $this->error('Command "openapi-generator" not found. Please install the OpenAPI Generator');
             $this->line('https://openapi-generator.tech/docs/installation/');
             return 1;
         }
 
-        $specPath = $this->buildSpecInTempPath($this->option('definition'));
+        if ($this->usingRedoc() && !$this->redocCommandExists()) {
+            $this->error('Command "redoc-cli" not found. Please install Redoc');
+            $this->line('https://www.npmjs.com/package/redoc-cli/');
+            return 1;
+        }
 
-        [$success, $output, $error] = $this->generate($this->argument('generator'), $specPath, $this->argument('output'));
+        [$generator, $outputPath, $specPath] = [
+            $this->argument('generator'),
+            $this->argument('output'),
+            $this->buildSpecInTempPath($this->option('definition')),
+        ];
+
+        if ($this->usingOpenApiGenerator()) {
+            [$success, $output, $error] = $this->generateWithOpenApiGenerator($generator, $specPath, $outputPath);
+        } elseif ($this->usingRedoc()) {
+            [$success, $output, $error] = $this->generateWithRedoc($specPath, $outputPath);
+        } else {
+            throw new \Exception;
+        }
 
         unlink($specPath);
 
@@ -41,7 +69,7 @@ class Generate extends \Illuminate\Console\Command
         return $specPath;
     }
 
-    public function generate(string $generator, string $specPath, string $output): array
+    public function generateWithOpenApiGenerator(string $generator, string $specPath, string $output): array
     {
         if ($jarPath = env('OPENAPI_GENERATOR_JAR_PATH')) {
             $javaHome = $this->getJavaHome();
@@ -65,7 +93,7 @@ class Generate extends \Illuminate\Console\Command
         ];
     }
 
-    private function openApiGeneratorCommandExists()
+    private function openApiGeneratorCommandExists(): bool
     {
         return (new Command('which openapi-generator'))->execute();
     }
@@ -77,5 +105,25 @@ class Generate extends \Illuminate\Console\Command
         $command->execute();
 
         return $command->getOutput();
+    }
+
+    private function generateWithRedoc(string $specPath, string $output): array
+    {
+        if (is_dir($output)) {$output .= '/index.html';}
+
+        $redocCommand = new Command('redoc-cli bundle '.$specPath);
+
+        $redocCommand->addArg('--output=', $output);
+
+        return [
+            $redocCommand->execute(),
+            $redocCommand->getOutput(),
+            $redocCommand->getError(),
+        ];
+    }
+
+    private function redocCommandExists(): bool
+    {
+        return (new Command('which redoc-cli'))->execute();
     }
 }
